blender_lib framework, and few gui changes.
The blender_lib framework is in the process of being framed out, along
with some basic gui.py changes.
Note that gui.py is in the middle of a change, which should have been on
its own commit. The next commit will be oriented towards basic
gui.py and blender_lib.gui code.