Added a small ctype change
I removed the get_type command from bcrf.ctypes, as it was rather
asinine heh. It has been placed in bcrf.blender_lib.object.Object.

Note that if there are other changes to object.py it is because a lot of
the framework is in the middle of a change, i just wanted to push the
ctype change.